# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Quiver API
  description: API for the Quiver surf community platform.
  version: 1.0.0
servers:
  - url: /api
    description: Local development / relative path
components:
  securitySchemes:
    supabaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token for authenticated user sessions.
    cronAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer token using CRON_SECRET for scheduled jobs. Also accepts x-vercel-cron header."
  schemas:
...
    ApiSuccess:
      type: object
      required: [success, timestamp, data]
      properties:
        success:
          type: boolean
          example: true
        timestamp:
          type: string
          format: date-time
        data:
          type: object
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    PaginationMeta:
      type: object
      properties:
        totalCount:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer

    # Core Resources
    Beach:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string, nullable: true }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
        country: { type: string, nullable: true }
        lat: { type: number }
        lon: { type: number }
        is_private: { type: boolean }
        created_at: { type: string, format: date-time }
...
paths:
  /beaches:
    get:
      summary: List all beaches
      responses:
        "200":
          description: List of beaches
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - properties:
                      data:
                        type: object
                        properties:
                          beaches:
                            type: array
                            items:
                              $ref: "#/components/schemas/Beach"
                          count:
                            type: integer
    post:
      summary: Create or update a beach (Admin only)
      security:
        - supabaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, lat, lon]
              properties:
                id:
                  {
                    type: string,
                    format: uuid,
                    description: "If provided, updates existing beach",
                  }
                name: { type: string }
                lat: { type: number }
                lon: { type: number }
      responses:
        "200":
          description: Beach created or updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - properties:
                      data:
                        type: object
                        properties:
                          success: { type: boolean }
                          data: { $ref: "#/components/schemas/Beach" }
...